// src/server/generated/question-answer-key.ts
export const BANK_VERSIONS = {
  "A": "871cf0290e5d89da926216f95eeecbc2ef46a5d8a10eaa2a29dab018736d14a3",
  "B": "492eaaa1e2d56b90923cd2ba13c901c0fc18e9c037db50462dafeb81b707a34b",
  "C": "94c525cdadf210ac3d183d17c80d9bef829eebcf1fe93e550917e5e39b5d0098"
} as const;
export const EXAM_RULES = {
  "A": {
    "total": 40,
    "single": 32,
    "multiple": 8,
    "pass": 30,
    "timeSeconds": 2400
  },
  "B": {
    "total": 60,
    "single": 45,
    "multiple": 15,
    "pass": 45,
    "timeSeconds": 3600
  },
  "C": {
    "total": 90,
    "single": 70,
    "multiple": 20,
    "pass": 70,
    "timeSeconds": 5400
  }
} as const;